Added handler to create field-config after adding a harvester. Corrected add-routine for field-config to find required fields by context.
